Add UI used to submit challenges (#40)
* Add Date Format utility methods and string resources

* Add Challenge Composable

* Add placeholder grid for future claim images

* Reformat code

* Change text padding

* Add tests

* Refactor code into composable sub-methods

* Handle plural and singular values


* Adapt to API changes and remove preview

* Add tests to check singular/plural hours/days formatting

* Change NEVER string to Never

* Remove old API

* Move components in correct packages

* Link challenge image with actual thumbnail

Changes:
 * Replace mock image with challenge.thumbnail
 * Implement and de-deprecate method getChallengeById from Database

* Modify imports

* Try (and fail) using database emulator for tests

* Migrate test toward the use of emulator

Changes:
 * Fix (possible) race condition (actually not because kotlin does not
   seems to have memory barrier)
 * Change the way we interract with firebase (using instance that can be
   modified)

* Fix test not working due to storage bad url

* Remove debug code

* Revert Text colour to Black

* Fix bad test (idk why this work)

Changes:
 * Modify the LazyRef so that it is no longer observable (use tasks
   instead)
 * Use waitUntil{} instead of future in task

---------
