+++ app/src/androidTest/java/com/github/geohunt/app/mocks/MockDatabase.kt
@@ -1,12 +1,12 @@
 package com.github.geohunt.app.mocks
 
 import android.graphics.Bitmap
+import com.github.geohunt.app.model.LazyRef
 import com.github.geohunt.app.model.database.Database
 import com.github.geohunt.app.model.database.api.Challenge
 import com.github.geohunt.app.model.database.api.Location
 import com.google.android.gms.tasks.Task
 import java.time.LocalDateTime
-import java.util.concurrent.CompletableFuture
 
 abstract class BaseMockDatabase : Database {
     override fun createChallenge(
@@ -17,7 +17,7 @@ abstract class BaseMockDatabase : Database {
         throw NotImplementedError()
     }
 
-    override fun getChallengeById(cid: String): Task<Challenge> {
+    override fun getChallengeById(cid: String): LazyRef<Challenge> {
         throw NotImplementedError()
     }
