+++ app/src/main/java/com/github/geohunt/app/model/database/Database.kt
@@ -31,8 +31,13 @@ interface Database {
         expirationDate: LocalDateTime? = null
     ): Task<Challenge>
 
-    @Deprecated("No longer used, users must privilege the use of LazyRef<Users> which does this in the background")
-    fun getChallengeById(cid: String): Task<Challenge>
+    /**
+     * Retrieve a challenge with a given ID and return a [LazyRef] upon completion
+     * 
+     * @param cid the challenge unique identifier
+     * @return A [LazyRef] linked to the result of the operation
+     */
+    fun getChallengeById(cid: String): LazyRef<Challenge>
 
     /**
      * Retrieve a list of challenges surrounding a particular location. Notice that the exact number of
